store proximity sightings

+++ firmware/nRF51/tag-proximity/inc/tracker.h
@@ -27,7 +27,7 @@
 
 extern const void* tracker_px(uint16_t listen_wait_ms);
 extern const void* tracker_tx(uint16_t listen_wait_ms, uint16_t tx_delay_us);
-extern void tracker_receive(uint32_t uid, int tx_power, int rx_power);
+extern void tracker_receive(uint32_t uid, int8_t tx_power, int8_t rx_power);
 extern void tracker_init(uint32_t uid);
 extern uint8_t tracker_loop(void);
 extern void tracker_second_tick(void);
